Fixing bugs in adding product to Shopping Card

+++ Blazor-E-Commerce.CoreBusiness/Models/OrderLineItem.cs
@@ -4,7 +4,7 @@
     {
         public int ID { get; set; }
         public int ProductID { get; set; }
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public int Quantity { get; set; }
         public int? OrderID { get; set; }
         public Product Product { get; set; } = null!;
